git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/ARQ/trunk@1122631 13f79535-47bb-0310-9956-ffa450edef68

+++ src/com/hp/hpl/jena/sparql/engine/ref/Eval.java
@@ -64,7 +64,7 @@ public class Eval
 
         DatasetGraph dsg = evaluator.getExecContext().getDataset() ;
         Iterator<Node> iter = dsg.listGraphNodes() ;
-        List<Binding> list = new ArrayList<Binding>(dsg.size()) ;
+        List<Binding> list = new ArrayList<Binding>((int)dsg.size()) ;
         for ( ; iter.hasNext(); )
         {
             Node gn = iter.next();
